Update on fibonacci 104: use long double

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,25 +6,25 @@
  * Return: 0 on success
  */
  
- int main(void)
- {
- 
- unsigned long int fib1 = 1, fib2 = 2, next_fib, count;
- 
- 	printf("%lu, %lu", fib1, fib2);
-	
+	int main(void)
+	{
+
+	long double fib1 = 1, fib2 = 2, next_fib, count;
+
+	printf("%lf, %lf", fib1, fib2);
+
 	for (count = 3; count <= 98; count++)
 	{
-	
+
 	next_fib = fib1 + fib2;
-	
-	printf(", %lu", next_fib);
-	
+
+	printf(", %lf", next_fib);
+
 	fib1 = fib2;
 	fib2 = next_fib;
 	}
-	
+
 	printf("\n");
-	
+
 	return 0;
 }
